Changes to remove Bliss/MACRO varargs hack.
In order to remove the Bliss/MACRO varargs hack that passed the VAX ap register
directly from the Pascal callers to the Pascal message printing functions, we
have to copy the $faol arguments into a small temp array to process them.
Taking the Address() of the arguments directly gives random results where
$faol prints the message string with random values or nothing for the params.

It was also necessary to copy the message string into a varying array of char
before passing it to $faol, because passing its Address() also doesn't work.

Updated ifcrtl_screen.pas to use system Pascal bindings instead of its own.
I'm going to update the rest of the game, but this file needed the most work.